chore(server): move event handlers to controllers
